https push

+++ cores/integrs/integr_httpserver.cpp
@@ -4,7 +4,7 @@
  *      Author: Lzy
  */
 #include "integr_httpserver.h"
-#include "file.h"
+#include "cfg_file.h"
 
 QPointer<JQHttpServer::Session> Integr_HttpServer::mSession;
 Integr_HttpServer::Integr_HttpServer(QObject *parent)
@@ -51,7 +51,7 @@ bool Integr_HttpServer::getDataItem(const QByteArray &body)
     Integr_JsonRecv *it = Integr_JsonRecv::bulid();
     bool ret = it->checkInput(body, object);
     if(ret) {
-        int res = it->getDataItem(object);
+        double res = it->getDataItem(object);
         replyValue(res);
     } else {
         ret = replyHttp("error", 212);
